asignar dispositivos listo

+++ src/app/Services/User/user.service.ts
@@ -32,41 +32,43 @@ export class UserService {
     console.log('busco device');
     this.deviceService.findDeviceById(device).subscribe((d) => {
       this.devicePro = d[0];
-      console.log('devicepro =device encontrado');
+      if (d[0]) {
+        console.log('devicepro =device encontrado');
+        this.db
+          .collection('users')
+          .doc(uid)
+          .get()
+          .subscribe((r) => {
+            console.log('me susbcribo a users');
 
-      this.db
-        .collection('users')
-        .doc(uid)
-        .get()
-        .subscribe((r) => {
-          console.log('me susbcribo a users');
+            if (r.data().devices) {
+              console.log('si hay data asigne');
 
-          if (r.data().devices) {
-            console.log('si hay data asigne');
+              this.listDeviceProv = r.data().devices;
+            } else {
+              console.log('si no hay data asigne []');
+              this.listDeviceProv = [];
+            }
+            this.listDeviceProv.push(this.devicePro);
+            console.table(this.listDeviceProv);
 
-            this.listDeviceProv = r.data().devices;
-          } else {
-            console.log('si no hay data asigne []');
-            this.listDeviceProv = [];
-          }
-          this.listDeviceProv.push(this.devicePro);
-          console.table(this.listDeviceProv);
-
-          this.db
-            .collection('users')
-            .doc(uid)
-            .update({
-              devices: this.listDeviceProv,
-            })
-            .then(() => {
-              console.log('Correcto');
-              console.log('borro documento');
-
-              this.deviceService.deleteDevice(d[0].uid);
-              console.log('documento borrado');
-            })
-            .catch((e) => console.log('error', e));
-        });
+            this.db
+              .collection('users')
+              .doc(uid)
+              .update({
+                devices: this.listDeviceProv,
+              })
+              .then(() => {
+                console.log('Correcto');
+                console.log('borro documento');
+                this.deviceService.deleteDevice(d[0].uid);
+                console.log('documento borrado');
+              })
+              .catch((e) => console.log('error', e));
+          });
+      } else {
+        console.log('No hay data');
+      }
     });
   }
